Move embeddings service out of enterprise
Mechanical code move to get another service out of enterprise.
