+++ WORKSPACE
@@ -404,7 +404,7 @@ load("//dev:oci_deps.bzl", "oci_deps")
 
 oci_deps()
 
-load("//enterprise/cmd/embeddings/shared:assets.bzl", "embbedings_assets_deps")
+load("//cmd/embeddings/shared:assets.bzl", "embbedings_assets_deps")
 
 embbedings_assets_deps()
