+++ cmd/embeddings/image_test.yaml
