+++ cmd/embeddings/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/sourcegraph/sourcegraph/enterprise/cmd/embeddings/shared"
+	"github.com/sourcegraph/sourcegraph/cmd/embeddings/shared"
 	"github.com/sourcegraph/sourcegraph/internal/sanitycheck"
 	"github.com/sourcegraph/sourcegraph/internal/service/svcmain"
 )
