+++ cmd/embeddings/qa/embed_queries/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 go_library(
     name = "embed_queries_lib",
     srcs = ["main.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/enterprise/cmd/embeddings/qa/embed_queries",
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/embeddings/qa/embed_queries",
     visibility = ["//visibility:private"],
     deps = [
         "//internal/conf",
