+++ cmd/embeddings/qa/embed_queries/main.go
