+++ cmd/embeddings/shared/assets.bzl
