+++ cmd/embeddings/shared/config.go
