+++ cmd/embeddings/shared/main.go
