+++ cmd/embeddings/shared/main_test.go
