+++ cmd/embeddings/shared/repo_embedding_index_cache.go
