+++ cmd/embeddings/shared/repo_embedding_index_cache_test.go
