+++ cmd/embeddings/shared/search.go
