+++ cmd/embeddings/shared/service.go
