+++ cmd/embeddings/shared/weaviate.go
