+++ cmd/sourcegraph/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//cmd/blobstore/shared",
+        "//cmd/embeddings/shared",
         "//cmd/executor/singlebinary",
         "//cmd/frontend/shared",
         "//cmd/gitserver/shared",
@@ -14,7 +15,6 @@ go_library(
         "//cmd/repo-updater/shared",
         "//cmd/searcher/shared",
         "//cmd/sourcegraph/osscmd",
-        "//enterprise/cmd/embeddings/shared",
         "//enterprise/cmd/symbols/shared",
         "//enterprise/cmd/worker/shared",
         "//internal/sanitycheck",
