+++ cmd/sourcegraph/main.go
@@ -10,13 +10,13 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/service/servegit"
 
 	blobstore_shared "github.com/sourcegraph/sourcegraph/cmd/blobstore/shared"
+	embeddings_shared "github.com/sourcegraph/sourcegraph/cmd/embeddings/shared"
 	executor_singlebinary "github.com/sourcegraph/sourcegraph/cmd/executor/singlebinary"
 	frontend_shared "github.com/sourcegraph/sourcegraph/cmd/frontend/shared"
 	gitserver_shared "github.com/sourcegraph/sourcegraph/cmd/gitserver/shared"
 	precise_code_intel_worker_shared "github.com/sourcegraph/sourcegraph/cmd/precise-code-intel-worker/shared"
 	repoupdater_shared "github.com/sourcegraph/sourcegraph/cmd/repo-updater/shared"
 	searcher_shared "github.com/sourcegraph/sourcegraph/cmd/searcher/shared"
-	embeddings_shared "github.com/sourcegraph/sourcegraph/enterprise/cmd/embeddings/shared"
 	symbols_shared "github.com/sourcegraph/sourcegraph/enterprise/cmd/symbols/shared"
 	worker_shared "github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/shared"
