+++ dev/check/go-dbconn-import.sh
@@ -20,7 +20,7 @@ allowed_prefix=(
   github.com/sourcegraph/sourcegraph/enterprise/cmd/migrator
   github.com/sourcegraph/sourcegraph/cmd/precise-code-intel-worker
   github.com/sourcegraph/sourcegraph/enterprise/cmd/symbols
-  github.com/sourcegraph/sourcegraph/enterprise/cmd/embeddings
+  github.com/sourcegraph/sourcegraph/cmd/embeddings
   # Doesn't connect but uses db internals for use with sqlite
   github.com/sourcegraph/sourcegraph/cmd/symbols
   # Transitively depends on zoekt package which imports but does not use DB
