+++ enterprise/dev/ci/internal/ci/bazel_operations.go
@@ -154,7 +154,7 @@ func bazelTest(targets ...string) func(*bk.Pipeline) {
 		bk.AllowDependencyFailure(),
 		bk.Agent("queue", "bazel"),
 		bk.Key("bazel-tests"),
-		bk.ArtifactPaths("./bazel-testlogs/enterprise/cmd/embeddings/shared/shared_test/*.log", "./command.profile.gz"),
+		bk.ArtifactPaths("./bazel-testlogs/cmd/embeddings/shared/shared_test/*.log", "./command.profile.gz"),
 		bk.AutomaticRetry(1), // TODO @jhchabran flaky stuff are breaking builds
 	}
